package/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rag"
version = "0.0.1"
description = "Local modular RAG engine"
requires-python = ">=3.9"

dependencies = [
    "sentence-transformers",
    "faiss-cpu"
]

[tool.setuptools]
packages = ["rag", "rag.engine"]
